Switched a failing spec to pending
Mongoid's *_changed? methods actually test the underlying attribute values;
whereas, ActiveModel's *_changed? methods guarantees to mark an attribute as
changed if the corresponding *_will_change! method was called. The result is
that it is difficult to mark a CarrierWave field as changed, because until
before_save is called, the file is only in the cache and not yet represented in
the attributes hash.

This might be related to issue #23